Security fixes toward Firefox 58 level but might miss some. Check later.
Security fixes toward Firefox 58 level but might miss some. Check later.

+++ js/src/frontend/TokenStream.h
@@ -1020,7 +1020,7 @@ class MOZ_STACK_CLASS TokenStream
         return true;
     }
 
-    void skipChars(uint8_t n) {
+    void skipChars(uint32_t n) {
         while (n-- > 0) {
             MOZ_ASSERT(userbuf.hasRawChars());
             mozilla::DebugOnly<int32_t> c = getCharIgnoreEOL();
